[treewide] lora: use int16_t for RSSI value
The RSSI values reported by LoRa transceiver can be less than -127.
Therefore, `int8_t` is not enough. This commit defines the RSSI of
`netdev_lora_rx_info` as `int16_t` and adapt the drivers accordingly
(sx126x, sx127x).

+++ pkg/semtech-loramac/contrib/semtech_loramac_radio.c
@@ -205,8 +205,8 @@ void SX127XStartCad(void)
 int16_t SX127XRssi(RadioModems_t modem)
 {
     (void)modem;
-    int8_t rssi;
-    loramac_netdev_ptr->driver->get(loramac_netdev_ptr, NETOPT_RSSI, &rssi, sizeof(int8_t));
+    int16_t rssi;
+    loramac_netdev_ptr->driver->get(loramac_netdev_ptr, NETOPT_RSSI, &rssi, sizeof(int16_t));
     return rssi;
 }
